RotatingFileHandler: Store complete path
Previously _created_files had been mixed. Sometimes the full path was stored and sometimes only the filename.

The problem with the filename only method was that there are cases where the logfile is not in the current working directory. Quill failed to rename that file in those cases. Basically quill::detail::rename_file() failed sometimes in _rotate_files() and showed an error on the console for files that did not exist.