Change the default open mode of FileSink to be 'a'
Change the default open mode of FileSink to be 'a' to make it match the comment.

The comment says the default open mode is 'a' but actually it is 'w'...

I think most people do not like the log file be overwritten to empty in every restart. Correct it be 'a' again.